Format + add delete confirm dialog

+++ src/app/services/cart.service.ts
@@ -10,6 +10,7 @@ export class CartService {
   private userId: number = 0;
   private carts: Cart[] = [];
   private currentCart?: Cart;
+
   constructor(
     private userService: UserService,
     private productService: ProductService
@@ -19,8 +20,6 @@ export class CartService {
   }
 
   public getCurrentCart(): Cart | undefined {
-  //let carts: Cart[] = JSON.parse(localStorage.getItem('carts') || '[]');
- // let currentCart: Cart | undefined = this.carts.find(cart => cart.userId === this.userId);
     this.currentCart = this.carts.find(cart => cart.userId === this.userId);
     return this.currentCart;
   }
@@ -33,18 +32,19 @@ export class CartService {
     }
     let productList: ProductInCart[] = [productInCart];
 
-    // Tạo biến cart và mảng carts
+    // Tạo biến cart
     let cart: Cart = {
       userId: this.userId,
       productList: productList
     }
+
     // Nếu carts trong localStorage chưa có -> Tạo key carts
     if (this.carts.length === 0) {
       this.carts = [cart]
       localStorage.setItem('carts', JSON.stringify(this.carts));
     } else {
       // Lấy ra cart của userId hiện tại
-      let currentCart: Cart | undefined = this.carts.find(cart => cart.userId === this.userId);
+      let currentCart: Cart | undefined = this.getCurrentCart();
 
       // Nếu tồn tại cart của userId hiện tại
       if (currentCart) {
@@ -66,34 +66,35 @@ export class CartService {
     }
   }
 
-  public changeQuantityProduct(productId: number, wantedQuantity: number):void{
-    let currentCart: Cart | undefined = this.carts.find(cart => cart.userId === this.userId);
+  public changeQuantityProduct(productId: number, wantedQuantity: number): void {
+    let currentCart = this.getCurrentCart();
     let currentProductInCart: ProductInCart | undefined =
-          currentCart?.productList.find(productInCart => productInCart.productId === productId);
+      currentCart?.productList.find(productInCart => productInCart.productId === productId);
     if (currentProductInCart) {
-        currentProductInCart.wantedQuantity = wantedQuantity
+      currentProductInCart.wantedQuantity = wantedQuantity
     }
     localStorage.setItem('carts', JSON.stringify(this.carts))
   }
 
   public removeProduct(productId: number): void {
-    this.currentCart?.productList.forEach((element,index)=>{
-      if(element.productId === productId) this.currentCart?.productList.splice(index,1);
+    if (!confirm('Bạn có muốn xóa?')) {
+      return;
+    }
+    this.currentCart?.productList.forEach((productInCart, index) => {
+      if (productInCart.productId === productId) this.currentCart?.productList.splice(index, 1);
     });
-   localStorage.setItem('carts', JSON.stringify(this.carts));
+    localStorage.setItem('carts', JSON.stringify(this.carts));
   }
 
   public getTotalPrice(): number {
-    var currentCart = this.getCurrentCart();
+    let currentCart = this.getCurrentCart();
     let totalPrice: number = 0;
-    if(currentCart){
-      currentCart?.productList.forEach((element : any) => {
-        let product = this.productService.getItemById(element.productId);
-        totalPrice += product.price * element.wantedQuantity;
+    if (currentCart) {
+      currentCart?.productList.forEach(productInCart => {
+        let product = this.productService.getItemById(productInCart.productId);
+        totalPrice += product.price * productInCart.wantedQuantity;
       });
     }
     return totalPrice;
   }
-
-
 }
